+++ tests/commands/Arc.test.js
@@ -93,11 +93,11 @@ describe('Arc', () => {
 				currentPoint
 			};
 
-			assert.equal(arc.export(settings, []), 'A10,10-45,0015,16');
+			assert.equal(arc.export(settings, []), 'A10,10-45 0015,16');
 			assert.equal(settings.currentPoint, new Point(15, 16));
 		});
 
-		it('should remove whitespace when compress=true and angle is a fraction', () => {
+		it('should not remove whitespace when compress=true and angle is a fraction', () => {
 			const arc = new Arc(['10,10 0.45 005,6']);
 			const settings = {
 				toAbsolute: true,
@@ -106,7 +106,20 @@ describe('Arc', () => {
 				currentPoint
 			};
 
-			assert.equal(arc.export(settings, []), 'A10,10.5,0015,16');
+			assert.equal(arc.export(settings, []), 'A10,10 .5 0015,16');
+			assert.equal(settings.currentPoint, new Point(15, 16));
+		});
+
+		it('should remove whitespace when compress=true and angle is a fraction', () => {
+			const arc = new Arc(['10,10.5 0.45 005,6']);
+			const settings = {
+				toAbsolute: true,
+				compress: true,
+				fractionDigits: 1,
+				currentPoint
+			};
+
+			assert.equal(arc.export(settings, []), 'A10,10.5.5 0015,16');
 			assert.equal(settings.currentPoint, new Point(15, 16));
 		});
 
@@ -118,7 +131,7 @@ describe('Arc', () => {
 				currentPoint
 			};
 
-			assert.equal(arc.export(settings, []), 'A10,10,45,0015,16');
+			assert.equal(arc.export(settings, []), 'A10,10 45 0015,16');
 			assert.equal(settings.currentPoint, new Point(15, 16));
 		});
 
@@ -130,7 +143,7 @@ describe('Arc', () => {
 				currentPoint
 			};
 
-			assert.equal(arc.export(settings, []), 'A10,10,0,0015,16');
+			assert.equal(arc.export(settings, []), 'A10,10 0 0015,16');
 			assert.equal(settings.currentPoint, new Point(15, 16));
 		});
 	});