+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html)
 
+## [0.4.1] - 2020-09-18
+### Fixed
+- Multiple consecutive shorthand lines now export correctly
+- Better logic around exporting fractions when compress = true
+
 ## [0.4.0] - 2020-09-17
 ### Fixed
 - Handling lack of whitespace around fractions and arc flags on import.